+++ src/Rest/FileProcessor.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Rest;
 
+use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Authorization\AuthorizationService;
 use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Entity\File;
 use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Service\DocumentService;
 use Dbp\Relay\CoreBundle\Rest\AbstractDataProcessor;
@@ -12,11 +13,10 @@ class FileProcessor extends AbstractDataProcessor
 {
     protected static string $identifierName = 'uid';
 
-    private DocumentService $documentService;
-
-    public function __construct(DocumentService $placeService)
+    public function __construct(
+        private readonly DocumentService $documentService,
+        private readonly AuthorizationService $authorizationService)
     {
-        $this->documentService = $placeService;
     }
 
     protected function addItem(mixed $data, array $filters): File
@@ -31,5 +31,10 @@ class FileProcessor extends AbstractDataProcessor
         assert($data instanceof File);
 
         return $this->documentService->replaceFile($identifier, $data);
+    }
+
+    protected function isCurrentUserGrantedOperationAccess(int $operation): bool
+    {
+        return $this->authorizationService->hasRoleUser();
     }
 }