+++ src/Rest/FileProvider.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Rest;
 
+use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Authorization\AuthorizationService;
 use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Entity\File;
 use Dbp\Relay\BlobConnectorCampusonlineDmsBundle\Service\DocumentService;
 use Dbp\Relay\CoreBundle\Rest\AbstractDataProvider;
@@ -15,11 +16,10 @@ class FileProvider extends AbstractDataProvider
 {
     protected static string $identifierName = 'uid';
 
-    private DocumentService $documentService;
-
-    public function __construct(DocumentService $placeService)
+    public function __construct(
+        private readonly DocumentService $documentService,
+        private readonly AuthorizationService $authorizationService)
     {
-        $this->documentService = $placeService;
     }
 
     protected function getItemById(string $id, array $filters = [], array $options = []): ?File
@@ -30,5 +30,10 @@ class FileProvider extends AbstractDataProvider
     protected function getPage(int $currentPageNumber, int $maxNumItemsPerPage, array $filters = [], array $options = []): array
     {
         throw new \RuntimeException('not available');
+    }
+
+    protected function isCurrentUserGrantedOperationAccess(int $operation): bool
+    {
+        return $this->authorizationService->hasRoleUser();
     }
 }