+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Add ROLE_USER policy to configure who is granted access to the API
+
 ## v0.1.2
 
 * Integrate Blob File API